requirements.txt
psutil
py-cpuinfo
distro